Refactor Swagger UI integration by removing the deprecated integration package and updating example implementations. Introduced a new Setup struct for better modularity and clarity in route registration.
