+++ _examples/server-sent-events/main.go
@@ -9,7 +9,6 @@ import (
 	"strconv"
 	"time"
 
-	"github.com/joakimcarlsson/go-router/integration"
 	"github.com/joakimcarlsson/go-router/openapi"
 	"github.com/joakimcarlsson/go-router/router"
 	"github.com/joakimcarlsson/go-router/swaggerui"
@@ -78,9 +77,9 @@ func main() {
 	uiConfig.Title = "SSE Examples"
 	uiConfig.DocExpansion = "list"
 
-	swaggerUI := integration.NewSwaggerUIIntegration(r, generator)
-	swaggerUI.WithUIConfig(uiConfig)
-	swaggerUI.SetupRoutes(r, "/openapi.json", "/docs")
+	setup := swaggerui.NewSetup(r, generator)
+	setup.WithUIConfig(uiConfig)
+	setup.RegisterRoutes(r, "/openapi.json", "/docs")
 
 	fmt.Println("Server starting on http://localhost:8080")
 	fmt.Println("API documentation available at http://localhost:8080/docs")
