--- integration/go.mod
@@ -1,15 +1,0 @@
-module github.com/joakimcarlsson/go-router/integration
-
-go 1.22
-
-require (
-	github.com/joakimcarlsson/go-router/openapi v0.0.0
-	github.com/joakimcarlsson/go-router/router v0.0.0
-	github.com/joakimcarlsson/go-router/swaggerui v0.0.0
-)
-
-replace (
-	github.com/joakimcarlsson/go-router/openapi => ../openapi
-	github.com/joakimcarlsson/go-router/router => ../router
-	github.com/joakimcarlsson/go-router/swaggerui => ../swaggerui
-)
