+++ swaggerui/go.mod
@@ -2,6 +2,12 @@ module github.com/joakimcarlsson/go-router/swaggerui
 
 go 1.22
 
-require github.com/joakimcarlsson/go-router/openapi v0.0.0
+require (
+	github.com/joakimcarlsson/go-router/openapi v0.0.0
+	github.com/joakimcarlsson/go-router/router v0.0.0
+)
 
-replace github.com/joakimcarlsson/go-router/openapi => ../openapi
+replace (
+	github.com/joakimcarlsson/go-router/openapi => ../openapi
+	github.com/joakimcarlsson/go-router/router => ../router
+)
